refactor(label): change file structure
