+++ packages/admin-ui/src/components/SearchBox/index.tsx
@@ -14,7 +14,7 @@ import { Button } from '../../button'
 import type { SystemComponent } from '../../types'
 import { useStateContext, StateContext } from './context'
 import * as style from './SearchBox.style'
-import { Label } from '../Label'
+import { Label } from '../../label'
 import { VisuallyHidden } from '../../visually-hidden'
 import type { Locale } from './intl'
 import { Intl, useLocale, LocaleProvider } from './intl'
