+++ packages/admin-ui/src/index.ts
@@ -43,8 +43,8 @@ export * from './divider'
 export * from './anchor'
 export * from './spinner'
 export * from './heading'
+export * from './label'
 
-export * from './components/Label'
 export * from './components/CollapsibleGroup'
 export * from './components/Avatar'
 export * from './components/Paragraph'
