+++ packages/admin-ui/src/label/index.ts
@@ -1,0 +1,1 @@
+export * from './label'
