+++ packages/admin-ui/src/label/label.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import type { Meta, Story } from '@storybook/react'
 
-import type { LabelProps } from './index'
-import { Label } from './index'
-import { Flex } from '../../flex'
+import type { LabelProps } from './label'
+import { Label } from './label'
+import { Flex } from '../flex'
 import { csx } from '@vtex/admin-ui-core'
 
 export default {
