+++ packages/admin-ui/src/switch/switch.style.ts
@@ -1,5 +1,5 @@
 import type { Theme } from '@vtex/admin-ui-core'
-import { focusVisible, style } from '@vtex/admin-ui-core'
+import { csx, focusVisible, style } from '@vtex/admin-ui-core'
 import { get } from '@vtex/admin-ui-util'
 
 // State styles
@@ -101,6 +101,6 @@ export const checked = style({
   ...disabled,
 })
 
-export const label = style({
+export const labelTheme = csx({
   cursor: 'pointer',
 })
